+++ src/reducers/filesReducer.js
@@ -2,12 +2,19 @@ import * as types from '../actions/ActionTypes'
 
 const filesReducer = (state=[], action) => {
   switch(action.type) {
-    case types.UPDATE_FILE:
-      return state.map(item => {
-        if(item.id !== action.file.id) return item
-        return Object.assign(item, action.file)
+    case types.ADD_FILE_SUCCESS:
+      return [
+        ...state.filter(file => file.guid !== action.file.guid),
+        Object.assign({}, action.file)
+      ]
+    case types.DELETE_FILE_SUCCESS:
+      return state.filter(file => file.guid !== action.file.guid)
+    case types.UPDATE_FILE_SUCCESS:
+      return state.map(file => {
+        if(file.guid !== action.file.guid) return file
+        return Object.assign(file, action.file)
       })
-    case types.REFRESH_FILES:
+    case types.LOAD_FILES_SUCCESS:
       return action.files
     default:
       return state