+++ src/reducers/index.test.js
@@ -2,17 +2,19 @@ import allReducers from '.'
 
 describe('Combine Reducers', () => {
   const initialState = {
-    filesReducer: [],
-    viewSwitcherReducer: false,
-    modalReducer: { modalType: null, modalProps: {} },
     breadcrumbReducer: [{ route: '/', libelle: 'Home' }],
+    filesReducer: [],
+    modalReducer: { modalType: null, modalProps: {} },
+    socketReducer: false,
+    viewSwitcherReducer: false,
   }
 
   it('should return the initial state', () => {
     const state = allReducers(undefined, {})
-    expect(state.filesReducer).toEqual(initialState.filesReducer)
-    expect(state.viewSwitcherReducer).toEqual(initialState.viewSwitcherReducer)
-    expect(state.modalReducer).toEqual(initialState.modalReducer)
     expect(state.breadcrumbReducer).toEqual(initialState.breadcrumbReducer)
+    expect(state.filesReducer).toEqual(initialState.filesReducer)
+    expect(state.modalReducer).toEqual(initialState.modalReducer)
+    expect(state.socketReducer).toEqual(initialState.socketReducer)
+    expect(state.viewSwitcherReducer).toEqual(initialState.viewSwitcherReducer)
   })
 })