+++ src/reducers/modalReducer.js
@@ -19,4 +19,4 @@ const modalReducer = (state=initialState, action) => {
   }
 }
 
-export default modalReducer;
+export default modalReducer