+++ src/store/configureStore.js
@@ -2,12 +2,11 @@ import { createStore, applyMiddleware } from 'redux'
 import thunk from 'redux-thunk'
 
 import allReducers from '../reducers'
-import socketMiddleware from '../middlewares/socketMiddleware'
 
 const configureStore = () => {
   return createStore(
     allReducers,
-    applyMiddleware(thunk, socketMiddleware)
+    applyMiddleware(thunk)
   )
 }